Game loop
I transplanted Rendering core's main method functionality over to the
game loop, so it initializes rendering core and calls a variation of the
run() method I made called render(), which does one frame at a time,
instead of code flow being caught in a loop over in run(). That looping
is handled by the game loop now. If we had more time, we could plug more
core functionality in, but ah well.